Re-wrote code snippet handling
Somehow the work I did last night stopped working. All I did was shut my laptop and open it up again today and run it. Either way, just added the snippet as an initialization parameter for the Questions constructor function, collected the divs in an array and used them in initializing the Question objects.